breaking out S3 RDDReader
this is payload agnostic
